update annotation and move knn related files to gts_common/components/retrieval_augmentations
