+++ gts_engine/gts_common/components/losses/__init__.py
@@ -1,4 +1,4 @@
-"""?
+"""loss计算工具集
 
 Todo:
     - [] (Jiang Yuzhen) 待修改和补全(详见components.__init__.py)
