+++ gts_engine/gts_common/components/metrics/__init__.py
@@ -1,4 +1,4 @@
-"""?
+"""评估工具集
 
 Todo:
     - [] (Jiang Yuzhen) 待修改和补全(详见components.__init__.py)
