+++ gts_engine/gts_common/components/retrieval_augmentations/knn_for_bagualu.py
@@ -1,4 +1,4 @@
-"""检索增强中的knn插值模块"""
+"""检索增强中的knn插值模块-适配八卦炉引擎"""
 import itertools
 import numpy as np
 from sklearn.metrics import pairwise_distances, accuracy_score
